requirements.txt
# AIQuant 量化交易系统 - 依赖包列表

# ===== 核心数据处理 =====
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# ===== 数据获取 =====
tushare>=1.4.0
akshare>=1.11.0

# ===== 数据库 =====
sqlalchemy>=2.0.0
pymysql>=1.1.0
psycopg2-binary>=2.9.0  # PostgreSQL（可选）

# ===== 数据存储 =====
h5py>=3.9.0
pyarrow>=13.0.0
redis>=5.0.0  # 缓存（可选）

# ===== 技术指标 =====
ta-lib>=0.4.28
pandas-ta>=0.3.14

# ===== 机器学习 =====
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0

# ===== 深度学习（可选）=====
# torch>=2.0.0
# tensorflow>=2.13.0

# ===== 回测框架 =====
backtrader>=1.9.78

# ===== 可视化 =====
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
mplfinance>=0.12.10

# ===== Web框架 =====
streamlit>=1.28.0  # 可视化面板
plotly>=5.17.0     # 交互式图表（已在上方声明）
# dash>=2.14.0     # 可选

# ===== 工具库 =====
joblib>=1.3.0
tqdm>=4.66.0
loguru>=0.7.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
click>=8.1.0
schedule>=1.2.0  # 任务调度
requests>=2.31.0  # HTTP请求

# ===== 测试 =====
pytest>=7.4.0
pytest-cov>=4.1.0

# ===== Jupyter =====
jupyter>=1.0.0
ipykernel>=6.25.0

